refactor : 불필요한 코드 및 주석 제거
